refactor: delete BackendObjType
`Arrow` and `AbstractArrow` become `GenArrow` and `GenAbstractArrow`, the last two
generated classes to leave the sealed hierarchy. `Native` is dissolved rather than
moved: with `BackendType.Reference` gone since #13138 its only remaining job was
`desc = clazz`, and `TypeDescs.toClassDesc` already returns `clazz` directly for
`SimpleType.Native`.

That empties the `BackendObjType` trait and the central `desc` match, so
**`BackendObjType.scala` is deleted** — the last of the old backend type layer.

`FunctionInterface` moves with `GenArrow`. Its `functionIns` now takes the argument
field it writes to instead of reaching for the enclosing `ArgField(0)`, which is
what all fifteen cases did.

Callers that held an `Arrow` or `AbstractArrow` value now hold the erased
`(args, result)` pair, or just the descriptor where that is all they used.
