+++ main/src/ca/uwaterloo/flix/language/phase/jvm/GenExpression.scala
@@ -23,7 +23,7 @@ import ca.uwaterloo.flix.language.ast.SemanticOp.*
 import ca.uwaterloo.flix.language.ast.shared.{Constant, ExpPosition, Mutability}
 import ca.uwaterloo.flix.language.ast.{SimpleType, *}
 import ca.uwaterloo.flix.language.phase.jvm.Instructions.*
-import ca.uwaterloo.flix.language.phase.jvm.classes.{GenCastError, GenEffectCall, GenExtTag, GenExtTagged, GenFrames, GenFramesNil, GenHandler, GenHoleError, GenLazy, GenMatchError, GenNullaryTag, GenRecord, GenRecordEmpty, GenRecordExtend, GenRegion, GenResult, GenResumption, GenResumptionNil, GenStruct, GenSuspension, GenTag, GenTagged, GenThunk, GenTuple, GenUnit, GenValue}
+import ca.uwaterloo.flix.language.phase.jvm.classes.{GenAbstractArrow, GenArrow, GenCastError, GenEffectCall, GenExtTag, GenExtTagged, GenFrames, GenFramesNil, GenHandler, GenHoleError, GenLazy, GenMatchError, GenNullaryTag, GenRecord, GenRecordEmpty, GenRecordExtend, GenRegion, GenResult, GenResumption, GenResumptionNil, GenStruct, GenSuspension, GenTag, GenTagged, GenThunk, GenTuple, GenUnit, GenValue}
 import ca.uwaterloo.flix.util.ClassDescs.internalNameOf
 import java.lang.constant.{ClassDesc, MethodTypeDesc}
 import java.lang.constant.ConstantDescs.{CD_double, CD_int, CD_long, CD_void}
@@ -1057,37 +1057,37 @@ object GenExpression {
 
     case Expr.ApplyClo(exp1, exp2, ct, _, purity, loc) =>
       // Type of the function abstract class
-      val functionInterface = BackendObjType.Arrow.fromArrowType(exp1.tpe)
-      val closureAbstractClass = BackendObjType.AbstractArrow.fromArrowType(exp1.tpe)
+      val (fnArgs, fnResult) = GenArrow.erasedArgsAndResult(exp1.tpe)
+      val closureAbstractClass = GenAbstractArrow.descOfArrowType(exp1.tpe)
       ct match {
         case ExpPosition.Tail =>
           // Evaluating the closure
           compileExpr(exp1)
           // Casting to JvmType of closure abstract class
-          mv.visitTypeInsn(Opcodes.CHECKCAST, internalNameOf(closureAbstractClass.desc))
+          mv.visitTypeInsn(Opcodes.CHECKCAST, internalNameOf(closureAbstractClass))
           // retrieving the unique thread object
-          mv.visitMethodInsn(Opcodes.INVOKEVIRTUAL, internalNameOf(closureAbstractClass.desc), closureAbstractClass.GetUniqueThreadClosureMethod.name, mkDescriptor()(closureAbstractClass.desc).descriptorString(), false)
+          mv.visitMethodInsn(Opcodes.INVOKEVIRTUAL, internalNameOf(closureAbstractClass), GenAbstractArrow.GetUniqueThreadClosureMethod(fnArgs, fnResult).name, mkDescriptor()(closureAbstractClass).descriptorString(), false)
           // Putting arg on the Fn class
           // Duplicate the FunctionInterface
           mv.visitInsn(Opcodes.DUP)
           // Evaluating the expression
           compileExpr(exp2)
-          PUTFIELD(functionInterface.ArgField(0))
+          PUTFIELD(GenArrow.ArgField(fnArgs, fnResult, 0))
           // Return the closure
           mv.visitInsn(Opcodes.ARETURN)
 
         case ExpPosition.NonTail =>
           compileExpr(exp1)
           // Casting to JvmType of closure abstract class
-          mv.visitTypeInsn(Opcodes.CHECKCAST, internalNameOf(closureAbstractClass.desc))
+          mv.visitTypeInsn(Opcodes.CHECKCAST, internalNameOf(closureAbstractClass))
           // retrieving the unique thread object
-          mv.visitMethodInsn(Opcodes.INVOKEVIRTUAL, internalNameOf(closureAbstractClass.desc), closureAbstractClass.GetUniqueThreadClosureMethod.name, mkDescriptor()(closureAbstractClass.desc).descriptorString(), false)
+          mv.visitMethodInsn(Opcodes.INVOKEVIRTUAL, internalNameOf(closureAbstractClass), GenAbstractArrow.GetUniqueThreadClosureMethod(fnArgs, fnResult).name, mkDescriptor()(closureAbstractClass).descriptorString(), false)
           // Putting arg on the Fn class
           // Duplicate the FunctionInterface
           mv.visitInsn(Opcodes.DUP)
           // Evaluating the expression
           compileExpr(exp2)
-          PUTFIELD(functionInterface.ArgField(0))
+          PUTFIELD(GenArrow.ArgField(fnArgs, fnResult, 0))
 
           // Calling unwind and unboxing
           if (Purity.isControlPure(purity)) {
@@ -1119,7 +1119,7 @@ object GenExpression {
       case ExpPosition.Tail =>
         val defInternalName = internalNameOf(GenFunAndClosureClasses.defnDesc(sym))
         // Type of the function abstract class
-        val functionInterface = BackendObjType.Arrow.fromArrowType(root.defs(sym).arrowType)
+        val (fnArgs, fnResult) = GenArrow.erasedArgsAndResult(root.defs(sym).arrowType)
 
         // Put the def on the stack
         mv.visitTypeInsn(Opcodes.NEW, defInternalName)
@@ -1131,7 +1131,7 @@ object GenExpression {
           mv.visitInsn(Opcodes.DUP)
           // Evaluating the expression
           compileExpr(arg)
-          PUTFIELD(functionInterface.ArgField(i))
+          PUTFIELD(GenArrow.ArgField(fnArgs, fnResult, i))
         }
         // Return the def
         mv.visitInsn(Opcodes.ARETURN)
@@ -1255,13 +1255,13 @@ object GenExpression {
     case Expr.ApplySelfTail(sym, exps, _, _, _) => ctx match {
       case EffectContext(_, _, _, setPc, _, _, _, _) =>
         // The function abstract class name
-        val functionInterface = BackendObjType.Arrow.fromArrowType(root.defs(sym).arrowType)
+        val (fnArgs, fnResult) = GenArrow.erasedArgsAndResult(root.defs(sym).arrowType)
         // Evaluate each argument and put the result on the Fn class.
         for ((arg, i) <- exps.zipWithIndex) {
           mv.visitVarInsn(Opcodes.ALOAD, 0)
           // Evaluate the argument and push the result on the stack.
           compileExpr(arg)
-          PUTFIELD(functionInterface.ArgField(i))
+          PUTFIELD(GenArrow.ArgField(fnArgs, fnResult, i))
         }
         mv.visitVarInsn(Opcodes.ALOAD, 0)
         pushInt(0)
@@ -1271,13 +1271,13 @@ object GenExpression {
 
       case DirectInstanceContext(_, _, _) =>
         // The function abstract class name
-        val functionInterface = BackendObjType.Arrow.fromArrowType(root.defs(sym).arrowType)
+        val (fnArgs, fnResult) = GenArrow.erasedArgsAndResult(root.defs(sym).arrowType)
         // Evaluate each argument and put the result on the Fn class.
         for ((arg, i) <- exps.zipWithIndex) {
           mv.visitVarInsn(Opcodes.ALOAD, 0)
           // Evaluate the argument and push the result on the stack.
           compileExpr(arg)
-          PUTFIELD(functionInterface.ArgField(i))
+          PUTFIELD(GenArrow.ArgField(fnArgs, fnResult, i))
         }
         // Jump to the entry point of the method.
         mv.visitJumpInsn(Opcodes.GOTO, ctx.entryPoint)
@@ -1523,7 +1523,7 @@ object GenExpression {
       for (HandlerRule(op, _, body) <- rules) {
         mv.visitInsn(Opcodes.DUP)
         compileExpr(body)
-        mv.visitFieldInsn(Opcodes.PUTFIELD, effectInternalName, GenEffectClasses.opName(op.sym), GenEffectClasses.opFieldType(op.sym).toDescriptor)
+        mv.visitFieldInsn(Opcodes.PUTFIELD, effectInternalName, GenEffectClasses.opName(op.sym), GenEffectClasses.opFieldType(op.sym).descriptorString())
       }
       // frames
       NEW(GenFramesNil.desc)
@@ -1584,7 +1584,7 @@ object GenExpression {
       methodExps.zipWithIndex.foreach { case (e, i) =>
         mv.visitInsn(Opcodes.DUP)
         compileExpr(e)
-        mv.visitFieldInsn(Opcodes.PUTFIELD, className, s"clo$i", BackendObjType.AbstractArrow.fromArrowType(e.tpe).toDescriptor)
+        mv.visitFieldInsn(Opcodes.PUTFIELD, className, s"clo$i", GenAbstractArrow.descOfArrowType(e.tpe).descriptorString())
       }
 
   }
