+++ main/src/ca/uwaterloo/flix/language/phase/jvm/classes/GenResumptionWrapper.scala
@@ -24,7 +24,7 @@ import ca.uwaterloo.flix.language.phase.jvm.ClassMaker.{ConstructorMethod, Const
 import ca.uwaterloo.flix.language.phase.jvm.Instructions.*
 import ca.uwaterloo.flix.language.phase.jvm.Mangle.{DevFlixRuntime, mkDesc}
 import ca.uwaterloo.flix.language.phase.jvm.MethodTypeDescs.mkDescriptor
-import ca.uwaterloo.flix.language.phase.jvm.{BackendObjType, Mangle, MethodTypeDescs}
+import ca.uwaterloo.flix.language.phase.jvm.{Mangle, MethodTypeDescs}
 import ca.uwaterloo.flix.util.ClassDescs
 import org.objectweb.asm.{Label, MethodVisitor, Opcodes}
 
@@ -43,11 +43,11 @@ object GenResumptionWrapper {
     mkDesc(DevFlixRuntime, Mangle.mkClassName("ResumptionWrapper", Mangle.erasedName(tpe)))
 
   // tpe -> Result
-  private def superClass(tpe: ClassDesc): BackendObjType.AbstractArrow =
-    BackendObjType.AbstractArrow(List(tpe), CD_Object)
+  private def superClassDesc(tpe: ClassDesc): ClassDesc =
+    GenAbstractArrow.desc(List(tpe), CD_Object)
 
   def genByteCode(tpe: ClassDesc)(implicit flix: Flix): Array[Byte] = {
-    val cm = mkClass(desc(tpe), IsFinal, superClass(tpe).desc)
+    val cm = mkClass(desc(tpe), IsFinal, superClassDesc(tpe))
     cm.mkConstructor(Constructor(tpe), IsPublic, constructorIns(tpe)(_))
     cm.mkField(ResumptionField(tpe), IsPrivate, IsFinal, NotVolatile)
     cm.mkMethod(Nil, InvokeMethod(tpe), IsPublic, NotFinal, invokeIns(tpe)(_))
@@ -60,7 +60,7 @@ object GenResumptionWrapper {
   private def constructorIns(tpe: ClassDesc)(implicit mv: MethodVisitor): Unit = {
     withName(1, GenResumption.desc) { resumption =>
       thisLoad()
-      INVOKESPECIAL(superClass(tpe).desc, ConstructorMethodName, MethodTypeDescs.NothingToVoid)
+      INVOKESPECIAL(superClassDesc(tpe), ConstructorMethodName, MethodTypeDescs.NothingToVoid)
       thisLoad()
       resumption.load()
       PUTFIELD(ResumptionField(tpe))
@@ -102,7 +102,7 @@ object GenResumptionWrapper {
   }
 
   private def UniqueMethod(tpe: ClassDesc): InstanceMethod =
-    InstanceMethod(desc(tpe), "getUniqueThreadClosure", mkDescriptor()(superClass(tpe).desc))
+    InstanceMethod(desc(tpe), "getUniqueThreadClosure", mkDescriptor()(superClassDesc(tpe)))
 
   private def uniqueIns(implicit mv: MethodVisitor): Unit = {
     thisLoad()
